Add files via upload
This is meant to be a factory for studentDBs where we can have multiple contracts keeping track of info for different sets of students, still a work in progress! not done yet, i need to figure out how to access student info directly via only name and contract index and nothing more!